backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.0.1
pydantic==2.11.7
requests==2.31.0
python-dotenv==1.0.0
rembg